practice-tasks/task4/requirements.txt
psycopg2-binary==2.9.9
mysql-connector-python==8.4.0
